+++ src/StraightScorer.Core/Services/Commands/MissCommand.cs
@@ -30,7 +30,7 @@ public class MissCommand(GameState _gameState): IUndoRedoCommand
 
         player.BreakCount++;
         player.BreakSum += player.CurrentBreak;
-        player.AverageBreak = player.BreakSum / player.BreakCount;
+        player.AverageBreak = (float)player.BreakSum / (float)player.BreakCount;
         player.HighestBreak = Math.Max(player.CurrentBreak, player.HighestBreak);
 
         // reset break and isAtTable
